[Model] Make product field 'price' -- BigDecimal

+++ src/test/java/ua/strongBody/dao/impl/BookingDAOImplUnitTest.java
@@ -15,6 +15,7 @@ import ua.strongBody.models.Cart;
 import ua.strongBody.models.Customer;
 import ua.strongBody.models.Product;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Collections;
 import java.util.List;
@@ -44,7 +45,7 @@ class BookingDAOImplUnitTest {
     private static final int PRODUCT_AMOUNT = 10;
     private static final String ORDER_NUMBER = "12345";
     private static final String PRODUCT_NAME = "prod1";
-    private static final int PRICE = 2000;
+    private static final BigDecimal PRICE = BigDecimal.valueOf(2000);
     private static final String ARTICLE = "PP-01";
     private static final String DESCRIPTION = "Super-puper product1";
     private static final int AVAILABLE_AMOUNT = 5;
